requirements.txt
Django==5.0.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
docker==6.0.1
boto3==1.26.1
gunicorn
pytest==7.3.2
pytest-django==4.5.2
python-dotenv==1.0.0
django-environ==0.12.0
drf-yasg
dj-rest-auth
django-allauth
django-rest-authtoken
django-extensions
Pillow
django-cors-headers
cryptography
python-decouple


